functions/health/requirements.txt
functions-framework==3.*
google-cloud-aiplatform==1.38.*
google-cloud-storage==2.10.*
google-cloud-secret-manager==2.16.*
google-cloud-logging==3.8.*
flask==2.3.*